adding resources & schedule components
