+++ client/src/components/profile/ProfilePageResources/ProfilePageResources.jsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import { resources } from '../../../util/resources';
+import { ButtonBubble } from '../../button/ButtonBubble/ButtonBubble';
+import './ProfilePageResources.scss';
+
+export const ProfilePageResources = () => {
+  return (
+    <div className="profile-page-resources profile-page-side-section">
+      <h2>Resources</h2>
+      {resources.map((resource, index) => {
+        return (
+          <a
+            key={index + resource.name}
+            href={resource.link}
+            target="_blank"
+            className="no-link-style"
+            rel="noreferrer"
+          >
+            <ButtonBubble
+              label={resource.name}
+              isSecondary
+              style={{ margin: 0, marginTop: '10px' }}
+            />
+          </a>
+        );
+      })}
+    </div>
+  );
+};
